requirements.txt
accelerate==1.10.1
av==15.1.0
beautifulsoup4
boto3
debugpy
depyf
diffusers
ffmpeg-python==0.2.0
ftfy
graphviz
imageio[ffmpeg]
loguru==0.7.3
mosaicml_streaming==0.8.0
packaging>=24.2
pandas
psycopg2-binary
pydantic
pydantic-settings
redis
redislite
rich
sentencepiece
setuptools>=78.1.1
timm==1.0.20
torchao
transformers==4.56.0
unfoldNd
versioningit
